fix: babysit round 2, 3 findings (3 coderabbit, 0 copilot, 0 ci)
- docs/reference/conventions.md:697 — fix tests/api/ → tests/unit/api/ in conftest scoping example (CodeRabbit).
- tests/e2e/{conftest.py,test_brownfield_intake_e2e.py} — move Windows pytest_asyncio_loop_factories hook from the test module to the tier conftest so pytest-asyncio discovers it via the normal plugin manager (CodeRabbit).
- tests/unit/api/conftest.py — rewrite loop-pinning rationale to drop the stale segfault claim; the SEGV did not reproduce under --count=2 (CodeRabbit; ADR 0006:388).
